sigh
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@949 0039d316-1c4b-4281-b951-d872f2087c98

+++ base/file_util_unittest.cc
@@ -27,6 +27,8 @@
 // (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
+#include "build/build_config.h"
+
 #if defined(OS_WIN)
 #include <windows.h>
 #include <shellapi.h>
@@ -89,7 +91,7 @@ class FindResultCollector {
   }
   
   int size() {
-    return files_.size();
+    return static_cast<int>(files_.size());
   }
 
  private:
